requirements.txt
Flask
psycopg2-binary 
ec2-metadata